requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dotenv==0.9.9
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mysql-connector-python==9.3.0
ollama==0.4.8
pip==24.3.1
pydantic==2.11.4
pydantic-core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
unicorn==2.1.3
urllib3==2.4.0
uvicorn==0.34.2
